move candidate queue into master package
